requirements.txt
fastapi
uvicorn
pyodbc
pydantic
python-dotenv
azure-identity